fix migration, manage model seeder

+++ database/migrations/2023_10_26_234603_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('price');
+            $table->bigInteger('price');
             $table->string('description');
             $table->foreignId('category_id')->constrained();
             $table->foreignId('user_id')->constrained();
